pyproject.toml

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "instamd"
version = "0.1.0"
dependencies = ["python-dotenv", "openai"]

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
"*" = ["*.json", "*.txt"]

[project.scripts]
instamd = "instamd.script:main"
